Move ECC.Crypto module to ecc subdirectory
